+++ cmake/sx126x_rf_api.cmake
@@ -2,21 +2,20 @@ include(ExternalProject)
 include(FetchContent)
 Set(FETCHCONTENT_QUIET FALSE)
 FetchContent_Declare(
-    addon_rfp
-    GIT_REPOSITORY "https://github.com/sigfox-tech-radio/sigfox-ep-addon-rfp"
-    GIT_TAG "v1.5"
+    sx126x_rf_api
+    GIT_REPOSITORY "https://github.com/sigfox-tech-radio/sigfox-ep-rf-api-semtech-sx126x"
+    GIT_TAG "v1.2"
     GIT_PROGRESS TRUE
     GIT_SHALLOW    1
     #SOURCE_DIR ${CMAKE_CURRENT_SOURCE_DIR}/addons/rfp
     UPDATE_DISCONNECTED TRUE
     STEP_TARGETS        update
 )
-FetchContent_GetProperties(addon_rfp)
+FetchContent_GetProperties(sx126x_rf_api)
 if (NOT platform_POPULATED)
-    FetchContent_Populate(addon_rfp)
-    add_subdirectory(${addon_rfp_SOURCE_DIR} ${addon_rfp_BINARY_DIR})
+    FetchContent_Populate(sx126x_rf_api)
+    add_subdirectory(${sx126x_rf_api_SOURCE_DIR} ${sx126x_rf_api_BINARY_DIR})
 endif()
 mark_as_advanced(FETCHCONTENT_QUIET FETCHCONTENT_BASE_DIR FETCHCONTENT_FULLY_DISCONNECTED FETCHCONTENT_UPDATES_DISCONNECTED)
-mark_as_advanced(FETCHCONTENT_SOURCE_DIR_ADDON_RFP FETCHCONTENT_UPDATES_DISCONNECTED_ADDON_RFP)
+mark_as_advanced(FETCHCONTENT_SOURCE_DIR_S_RF_API FETCHCONTENT_UPDATES_DISCONNECTED_SX126X_RF_API)
 #FetchContent_MakeAvailable(addon_rfp)
-